+++ data/src/main/java/com/fernandocejas/android10/sample/data/entity/weather/Clouds.java
@@ -1,9 +1,5 @@
-
 package com.fernandocejas.android10.sample.data.entity.weather;
 
-
 public class Clouds {
-
-    public Integer all;
-
+  public float all;
 }
